requirements.txt
joblib==1.1.0
numpy==1.19.5
scikit_learn==1.0.1
tensorflow_addons==0.15.0
tensorflow_gpu==2.5.0
